Move normalization helpers
