+++ src/backend/tfjs_backend.ts
@@ -351,106 +351,6 @@ export function sliceAlongAxis(
   });
 }
 
-
-/**
- * Non-broadcasting batch normalization for use in training (not inference).
- *
- * The input is normalized to zero mean and unit variance along the
- * `reductionAxes`, followed by scaling with `gamma` and shifted by `beta`.
- * The result of that is returned as the first element
- * of the returned `Array`. The other two elements are the mean and variance,
- * respectively.
- *
- * @param x Input tensor to be normalized.
- * @param gamma Tensor by which to scale the input.
- * @param beta Tensor by which to center the input.
- * @param reductionAxes Axes over which to normalize.
- * @param epsilon Fuzz factor.
- * @returns An `Array` of three `Tensors`:
- *   [normalized tensor, mean of input, variance of input].
- */
-function regularNormalizeBatchInTraining(
-    x: Tensor, gamma: Tensor, beta: Tensor, reductionAxes: number[],
-    epsilon = 1e-3): [Tensor, Tensor, Tensor] {
-  return tidy(() => {
-           const meanAndVariance = tfc.moments(x, reductionAxes);
-           const mean = meanAndVariance.mean;
-           const variance = meanAndVariance.variance;
-           const normed =
-               batchNormalization(x, mean, variance, beta, gamma, epsilon);
-           return [normed, mean, variance];
-         }) as [Tensor, Tensor, Tensor];
-}
-
-/**
- * Broadcasting batch normalization for use in training (not inference).
- *
- * The input is normalized to zero mean and unit variance along the
- * `reductionAxes`, followed by scaling with `gamma` and shifted by `beta`.
- * The result of that is returned as the first element
- * of the returned `Array`. The other two elements are the mean and variance,
- * respectively.
- *
- * @param x Input tensor to be normalized.
- * @param gamma Tensor by which to scale the input.
- * @param beta Tensor by which to center the input.
- * @param reductionAxes Axes over which to normalize.
- * @param epsilon Fuzz factor.
- * @returns An `Array` of three `Tensors`:
- *   [normalized tensor, mean of input, variance of input].
- */
-function broadcastNormalizeBatchInTraining(
-    x: Tensor, gamma: Tensor, beta: Tensor, reductionAxes: number[],
-    epsilon = 1e-3): [Tensor, Tensor, Tensor] {
-  return tidy(() => {
-           const meanAndVariance = tfc.moments(x, reductionAxes);
-           const mean = meanAndVariance.mean;
-           const variance = meanAndVariance.variance;
-           const targetShape: number[] = [];
-           for (const axis of math_utils.range(0, ndim(x))) {
-             if (reductionAxes.indexOf(axis) !== -1) {
-               targetShape.push(1);
-             } else {
-               targetShape.push(x.shape[axis]);
-             }
-           }
-           const broadcastMean = mean.reshape(targetShape);
-           const broadcastVariance = variance.reshape(targetShape);
-           const broadcastGamma =
-               gamma == null ? null : gamma.reshape(targetShape);
-           const broadcastBeta =
-               beta == null ? null : beta.reshape(targetShape);
-           const normed = batchNormalization(
-               x, broadcastMean, broadcastVariance, broadcastBeta,
-               broadcastGamma, epsilon);
-           return [normed, mean, variance];
-         }) as [Tensor, Tensor, Tensor];
-}
-
-/**
- * Batch normalization for use in training (not inference).
- *
- * @param x Input tensor to be normalized.
- * @param gamma Tensor by which to scale the input.
- * @param beta Tensor by which to center the input.
- * @param reductionAxes Axes over which to normalize.
- * @param epsilon Fuzz factor.
- * @returns An `Array` of three `Tensors`:
- *   [normalized tensor, mean of input, variance of input].
- */
-export function normalizeBatchInTraining(
-    x: Tensor, gamma: Tensor, beta: Tensor, reductionAxes: number[],
-    epsilon = 1e-3): [Tensor, Tensor, Tensor] {
-  if (util.arraysEqual(
-          reductionAxes.slice().sort(), math_utils.range(0, ndim(x) - 1))) {
-    return regularNormalizeBatchInTraining(
-        x, gamma, beta, reductionAxes, epsilon);
-  } else {
-    return broadcastNormalizeBatchInTraining(
-        x, gamma, beta, reductionAxes, epsilon);
-  }
-}
-
 /**
  * Concatenates a list of tensors alongside the specified axis.
  * @param tensors `Array` of tensors to concatenate.
@@ -810,50 +710,6 @@ export function pow(x: Tensor, a: Tensor|number): Tensor {
     }
     return tfc.pow(x, a as Tensor);
   });
-}
-
-/* Normalization operations. */
-
-/**
- * Applies batch normalization on x given mean, var, beta and gamma.
- *
- * I.e. returns:
- *   `output = (x - mean) / (sqrt(var) + epsilon) * gamma + beta`
- *
- * @param x Input tensor.
- * @param mean Mean of batch.
- * @param variance Variance of batch.
- * @param beta Tensor with which to center the input.
- * @param gamma Tensor by which to scale the input.
- * @param epsilon Fuzz factor.
- * @returns The result of the batch normalization.
- */
-export function batchNormalization(
-    x: Tensor, mean: Tensor, variance: Tensor, beta?: Tensor, gamma?: Tensor,
-    epsilon = 1e-3): Tensor {
-  let out: Tensor;
-  if (ndim(x) === 2) {
-    out = tfc.batchNormalization2d(
-        x as Tensor2D, mean as Tensor2D | Tensor1D,
-        variance as Tensor2D | Tensor1D, epsilon, gamma as Tensor2D | Tensor1D,
-        beta as Tensor2D | Tensor1D);
-  } else if (ndim(x) === 3) {
-    // TODO(cais): Check rank; give proper error message.
-    out = tfc.batchNormalization3d(
-        x as Tensor3D, mean as Tensor3D | Tensor1D,
-        variance as Tensor3D | Tensor1D, epsilon, gamma as Tensor3D | Tensor1D,
-        beta as Tensor3D | Tensor1D);
-  } else if (ndim(x) === 4) {
-    out = tfc.batchNormalization4d(
-        x as Tensor4D, mean as Tensor4D | Tensor1D,
-        variance as Tensor4D | Tensor1D, epsilon, gamma as Tensor4D | Tensor1D,
-        beta as Tensor4D | Tensor1D);
-  } else {
-    throw new NotImplementedError(
-        `batchNormalization is not implememnted for array of rank ${ndim(x)} ` +
-        `yet`);
-  }
-  return out;
 }
 
 /* Neural-network operations. */
