+++ src/layers/normalization_test.ts
@@ -13,11 +13,13 @@
  */
 
 // tslint:disable:max-line-length
-import {onesLike, Tensor, tensor2d, tensor3d, tensor4d, train, zeros, zerosLike} from '@tensorflow/tfjs-core';
+import {onesLike, Tensor, tensor1d, tensor2d, tensor3d, tensor4d, train, zeros, zerosLike} from '@tensorflow/tfjs-core';
 
 import * as tfl from '../index';
 import {SymbolicTensor} from '../types';
 import {describeMathCPU, describeMathCPUAndGPU, expectTensorsClose} from '../utils/test_utils';
+import {batchNormalization, normalizeBatchInTraining} from './normalization';
+
 // tslint:enable:max-line-length
 
 describeMathCPU('BatchNormalization Layers: Symbolic', () => {
@@ -282,5 +284,277 @@ describeMathCPUAndGPU('BatchNormalization Layers: Tensor', () => {
         .catch(err => {
           console.error(err.stack);
         });
+  });
+});
+
+describeMathCPUAndGPU('normalizeBatchInTraining', () => {
+  // The reference values for assertion below can be obtained with Python code
+  // as the following:
+  // ```python
+  // import keras
+  // import numpy as np
+  // import tensorflow as tf
+  //
+  // with tf.Session() as sess:
+  //   x = tf.Variable(np.array(
+  //       [[1, 2, 3, 4], [2, 4, 6, 8], [12, 11, 10, 9]], dtype=np.float32))
+  //   gamma = tf.Variable(np.array([1, 1, 1, 1], dtype=np.float32))
+  //   beta = tf.Variable(np.array([0, 0, 0, 0], dtype=np.float32))
+  //   reduction_axes = [0]
+  //   normed, mean, variance = keras.backend.normalize_batch_in_training(
+  //       x, gamma, beta, reduction_axes)
+  //   print(normed)
+  //   print(mean)
+  //   print(variance)
+  // ```
+
+  it('2D, no broadcasting', () => {
+    const x = tensor2d([[1, 2, 3, 4], [2, 4, 6, 8], [12, 11, 10, 9]], [3, 4]);
+    const gamma = tensor1d([1, 1, 1, 1]);
+    const beta = tensor1d([0, 0, 0, 0]);
+    const reductionAxes = [0];
+    const [normed, mean, variance] =
+        normalizeBatchInTraining(x, gamma, beta, reductionAxes);
+    expectTensorsClose(
+        normed,
+        tensor2d(
+            [
+              [-0.805371, -0.9502233, -1.1624058, -1.3885813],
+              [-0.6040282, -0.4319197, -0.11624074, 0.46286058],
+              [1.4093992, 1.3821429, 1.2786462, 0.92572117]
+            ],
+            [3, 4]));
+    expectTensorsClose(mean, tensor1d([5.0, 5.6666665, 6.3333335, 7.0]));
+    expectTensorsClose(
+        variance, tensor1d([24.666666, 14.888889, 8.222222, 4.6666665]));
+  });
+
+  it('3D, no broadcasting', () => {
+    const x = tensor3d(
+        [[[1, 2], [3, 4]], [[2, 4], [6, 8]], [[12, 11], [10, 9]]], [3, 2, 2]);
+    const gamma = tensor1d([1, 1]);
+    const beta = tensor1d([0, 0]);
+    const reductionAxes = [0, 1];
+    const [normed, mean, variance] =
+        normalizeBatchInTraining(x, gamma, beta, reductionAxes);
+    expectTensorsClose(
+        normed,
+        tensor3d(
+            [
+              [[-1.1355163, -1.3552775], [-0.6488664, -0.7297648]],
+              [[-0.8921913, -0.7297648], [0.08110833, 0.5212605]],
+              [[1.5410578, 1.4595294], [1.0544081, 0.8340168]]
+            ],
+            [3, 2, 2]));
+    expectTensorsClose(mean, tensor1d([5.6666665, 6.3333335]));
+    expectTensorsClose(variance, tensor1d([16.88889, 10.222222]));
+  });
+
+  it('3D, broadcasting', () => {
+    const x = tensor3d(
+        [[[1, 2], [3, 4]], [[2, 4], [6, 8]], [[12, 11], [10, 9]]], [3, 2, 2]);
+    const gamma = tensor2d([[1, 1], [1, 1]], [2, 2]);
+    const beta = tensor2d([[0, 0], [0, 0]], [2, 2]);
+    const reductionAxes = [0];
+    const [normed, mean, variance] =
+        normalizeBatchInTraining(x, gamma, beta, reductionAxes);
+    expectTensorsClose(
+        normed,
+        tensor3d(
+            [
+              [[-0.805371, -0.9502233], [-1.1624058, -1.3885813]],
+              [[-0.6040282, -0.4319197], [-0.11624074, 0.46286058]],
+              [[1.4093992, 1.3821429], [1.2786462, 0.92572117]]
+            ],
+            [3, 2, 2]));
+    expectTensorsClose(
+        mean, tensor2d([[5, 5.6666665], [6.3333335, 7]], [2, 2]));
+    expectTensorsClose(
+        variance,
+        tensor2d([[24.666666, 14.888889], [8.222222, 4.6666665]], [2, 2]));
+  });
+
+  it('4D, broadcasting', () => {
+    const x = tensor4d(
+        [[[[1, 2], [3, 4]], [[2, 4], [6, 8]], [[12, 11], [10, 9]]]],
+        [1, 3, 2, 2]);
+    const gamma = tensor2d([[1, 1], [1, 1]], [2, 2]);
+    const beta = tensor2d([[0, 0], [0, 0]], [2, 2]);
+    const reductionAxes = [0, 1];
+    const [normed, mean, variance] =
+        normalizeBatchInTraining(x, gamma, beta, reductionAxes);
+    expectTensorsClose(
+        normed,
+        tensor4d(
+            [[
+              [[-0.805371, -0.9502233], [-1.1624058, -1.3885813]],
+              [[-0.6040282, -0.4319197], [-0.11624074, 0.46286058]],
+              [[1.4093992, 1.3821429], [1.2786462, 0.92572117]]
+            ]],
+            [1, 3, 2, 2]));
+    expectTensorsClose(
+        mean, tensor2d([[5, 5.6666665], [6.3333335, 7]], [2, 2]));
+    expectTensorsClose(
+        variance,
+        tensor2d([[24.666666, 14.888889], [8.222222, 4.6666665]], [2, 2]));
+  });
+});
+
+describeMathCPUAndGPU('batchNormalization', () => {
+  it('2D, no broadcast, no gamma, no beta', () => {
+    const x = tensor2d([[10, 20], [30, 40]], [2, 2]);
+    const mean = tensor2d([[5, 5], [5, 5]], [2, 2]);
+    const variance = tensor2d([[4, 16], [4, 16]], [2, 2]);
+    expectTensorsClose(
+        batchNormalization(x, mean, variance, null, null, 0),
+        tensor2d([[2.5, 3.75], [12.5, 8.75]], [2, 2]));
+  });
+  it('2D, no broadcast, no gamma, no beta, custom epsilon', () => {
+    const x = tensor2d([[30, 30], [60, 60]], [2, 2]);
+    const mean = tensor2d([[0, 0], [0, 0]], [2, 2]);
+    const variance = tensor2d([[7, 7], [7, 7]], [2, 2]);
+    expectTensorsClose(
+        batchNormalization(x, mean, variance, null, null, 2),
+        tensor2d([[10, 10], [20, 20]], [2, 2]));
+  });
+  it('2D, no broadcast, gamma, no beta', () => {
+    const x = tensor2d([[10, 20], [30, 40]], [2, 2]);
+    const mean = tensor2d([[5, 5], [5, 5]], [2, 2]);
+    const variance = tensor2d([[4, 16], [4, 16]], [2, 2]);
+    const gamma = tensor2d([[1, 2], [3, 4]], [2, 2]);
+    expectTensorsClose(
+        batchNormalization(x, mean, variance, null, gamma, 0),
+        tensor2d([[2.5, 7.5], [37.5, 35]], [2, 2]));
+  });
+  it('2D, no broadcast, gamma, beta', () => {
+    const x = tensor2d([[10, 20], [30, 40]], [2, 2]);
+    const mean = tensor2d([[5, 5], [5, 5]], [2, 2]);
+    const variance = tensor2d([[4, 16], [4, 16]], [2, 2]);
+    const gamma = tensor2d([[1, 2], [3, 4]], [2, 2]);
+    const beta = tensor2d([[-1, -1], [-2, -2]], [2, 2]);
+    expectTensorsClose(
+        batchNormalization(x, mean, variance, beta, gamma, 0),
+        tensor2d([[1.5, 6.5], [35.5, 33]], [2, 2]));
+  });
+  it('2D, broadcast, gamma, beta', () => {
+    const x = tensor2d([[10, 20], [30, 40]], [2, 2]);
+    const mean = tensor1d([2, 5]);
+    const variance = tensor1d([1, 4]);
+    const gamma = tensor1d([3, 4]);
+    const beta = tensor1d([-1, -2]);
+    expectTensorsClose(
+        batchNormalization(x, mean, variance, beta, gamma, 0),
+        tensor2d([[23, 28], [83, 68]], [2, 2]));
+  });
+  it('3D, no broadcast, no gamma, no beta', () => {
+    const x = tensor3d([[[10, 20], [30, 40]], [[10, 20], [30, 40]]], [2, 2, 2]);
+    const mean = tensor3d([[[5, 5], [5, 5]], [[5, 5], [5, 5]]], [2, 2, 2]);
+    const variance =
+        tensor3d([[[4, 16], [4, 16]], [[16, 25], [16, 25]]], [2, 2, 2]);
+    expectTensorsClose(
+        batchNormalization(x, mean, variance, null, null, 0),
+        tensor3d(
+            [[[2.5, 3.75], [12.5, 8.75]], [[1.25, 3], [6.25, 7]]], [2, 2, 2]));
+  });
+  it('3D, no broadcast, gamma, beta', () => {
+    const x = tensor3d([[[10, 20], [30, 40]], [[10, 20], [30, 40]]], [2, 2, 2]);
+    const mean = tensor3d([[[5, 5], [5, 5]], [[5, 5], [5, 5]]], [2, 2, 2]);
+    const variance =
+        tensor3d([[[4, 16], [4, 16]], [[16, 25], [16, 25]]], [2, 2, 2]);
+    const gamma = tensor3d([[[2, 2], [2, 2]], [[4, 4], [4, 4]]], [2, 2, 2]);
+    const beta =
+        tensor3d([[[-1, -1], [-2, -2]], [[-1, -1], [-2, -2]]], [2, 2, 2]);
+    expectTensorsClose(
+        batchNormalization(x, mean, variance, beta, gamma, 0),
+        tensor3d([[[4, 6.5], [23, 15.5]], [[4, 11], [23, 26]]], [2, 2, 2]));
+  });
+  it('3D, broadcast, gamma, beta', () => {
+    const x = tensor3d([[[10, 20], [30, 40]], [[10, 20], [30, 40]]], [2, 2, 2]);
+    const mean = tensor1d([5, 5]);
+    const variance = tensor1d([4, 16]);
+    const gamma = tensor1d([2, 4]);
+    const beta = tensor1d([-1, -2]);
+    expectTensorsClose(
+        batchNormalization(x, mean, variance, beta, gamma, 0),
+        tensor3d([[[4, 13], [24, 33]], [[4, 13], [24, 33]]], [2, 2, 2]));
+  });
+  it('4D, no broadcast, no gamma, no beta', () => {
+    const x = tensor4d(
+        [
+          [[[10, 20], [30, 40]], [[10, 20], [30, 40]]],
+          [[[-10, -20], [-30, -40]], [[-10, -20], [-30, -40]]]
+        ],
+        [2, 2, 2, 2]);
+    const mean = tensor4d(
+        [
+          [[[5, 5], [5, 5]], [[5, 5], [5, 5]]],
+          [[[-5, -5], [-5, -5]], [[-5, -5], [-5, -5]]]
+        ],
+        [2, 2, 2, 2]);
+    const variance = tensor4d(
+        [
+          [[[4, 16], [4, 16]], [[16, 25], [16, 25]]],
+          [[[4, 16], [4, 16]], [[16, 25], [16, 25]]]
+        ],
+        [2, 2, 2, 2]);
+    expectTensorsClose(
+        batchNormalization(x, mean, variance, null, null, 0),
+        tensor4d(
+            [
+              [[[2.5, 3.75], [12.5, 8.75]], [[1.25, 3], [6.25, 7]]],
+              [[[-2.5, -3.75], [-12.5, -8.75]], [[-1.25, -3], [-6.25, -7]]]
+            ],
+            [2, 2, 2, 2]));
+  });
+  it('4D, no broadcast, gamma, beta', () => {
+    const x = tensor4d(
+        [
+          [[[10, 20], [30, 40]], [[10, 20], [30, 40]]],
+          [[[-10, -20], [-30, -40]], [[-10, -20], [-30, -40]]]
+        ],
+        [2, 2, 2, 2]);
+    const mean = tensor4d(
+        [
+          [[[5, 5], [5, 5]], [[5, 5], [5, 5]]],
+          [[[-5, -5], [-5, -5]], [[-5, -5], [-5, -5]]]
+        ],
+        [2, 2, 2, 2]);
+    const variance = tensor4d(
+        [
+          [[[4, 16], [4, 16]], [[16, 25], [16, 25]]],
+          [[[4, 16], [4, 16]], [[16, 25], [16, 25]]]
+        ],
+        [2, 2, 2, 2]);
+    const gamma = tensor4d(
+        [
+          [[[2, 2], [2, 2]], [[4, 4], [4, 4]]],
+          [[[2, 2], [2, 2]], [[4, 4], [4, 4]]]
+        ],
+        [2, 2, 2, 2]);
+    const beta = tensor4d(
+        [
+          [[[-1, -1], [-2, -2]], [[-1, -1], [-2, -2]]],
+          [[[1, 1], [2, 2]], [[1, 1], [2, 2]]]
+        ],
+        [2, 2, 2, 2]);
+    expectTensorsClose(
+        batchNormalization(x, mean, variance, beta, gamma, 0),
+        tensor4d(
+            [
+              [[[4, 6.5], [23, 15.5]], [[4, 11], [23, 26]]],
+              [[[-4, -6.5], [-23, -15.5]], [[-4, -11], [-23, -26]]]
+            ],
+            [2, 2, 2, 2]));
+  });
+  it('4D, broadcast, gamma, beta', () => {
+    const x = tensor4d(
+        [[[[10, 20], [30, 40]]], [[[10, 20], [30, 40]]]], [2, 1, 2, 2]);
+    const mean = tensor1d([5, 5]);
+    const variance = tensor1d([4, 16]);
+    const gamma = tensor1d([2, 4]);
+    const beta = tensor1d([-1, -2]);
+    expectTensorsClose(
+        batchNormalization(x, mean, variance, beta, gamma, 0),
+        tensor4d([[[[4, 13], [24, 33]]], [[[4, 13], [24, 33]]]], [2, 1, 2, 2]));
   });
 });
